Replace EventBus and EventEmitter with a single class named EventEmitterImpl
